update logic for delete methods

+++ src/CommerceApi.DataAccessLayer/ApplicationDataContext.cs
@@ -56,14 +56,24 @@ public partial class ApplicationDataContext : AuthenticationDataContext, IDataCo
         ArgumentNullException.ThrowIfNull(entity, nameof(entity));
 
         var set = Set<TEntity>();
+        if (entity is DeletableEntity)
+        {
+            set.Attach(entity);
+        }
+
         set.Remove(entity);
     }
 
     public void Delete<TEntity>(IEnumerable<TEntity> entities) where TEntity : BaseEntity
     {
-        ArgumentNullException.ThrowIfNull(entities, nameof(entities));
+        var hasItems = entities?.Any() ?? false;
+        if (!hasItems)
+        {
+            throw new ArgumentNullException(nameof(entities), "you must provide at least one entity");
+        }
 
         var set = Set<TEntity>();
+        //set.AttachRange(entities);
         set.RemoveRange(entities);
     }
